feature: tratamento de erros

+++ src/main/java/com/api/izacar/repository/MarcaRepository.java
@@ -1,7 +1,7 @@
 package com.api.izacar.repository;
 
-import com.api.izacar.models.Marca;
+import com.api.izacar.models.marca.Marca;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MarcaRepository extends JpaRepository<Marca, Long> {
+public interface MarcaRepository extends JpaRepository<Marca, Integer> {
 }
